re-organize file utilities into file_utils, minor refactor with tests, main.rs is clean again
